Added a helper class for handling photos.
Tried adding a unit test for this new implementation, but was unsure about
the behavior when working with the filesystem.

Will add these unit tests as bugs are reported and fixed.

The photograph class is well documented and should be read for completeness.

The setImageUrl() method was overridden in the Recipe class to create the
thumbnail, but otherwise nothing should have changed.